User changes and migration changes

+++ db/migrate/20130103065345_change_login_system.rb
@@ -1,13 +1,13 @@
 class User < ActiveRecord::Base
-  attr_accessible :password, :profile
-  attr_accessor :password
-  before_save :encrypt_password
-  belongs_to :profile, polymorphic: true
-  validates :email, presence: true, format: { with: /.+@.+\..+/ }
+  attr_accessible :profile
 
-  def encrypt_password
-    self.password_salt = BCrypt::Engine.generate_salt
-    self.password_hash = BCrypt::Engine.hash_secret(password, password_salt)
+  belongs_to :profile, polymorphic: true
+  validates :email, format: { with: /.+@.+\..+/ }, allow_blank: true
+
+  before_save :set_salt
+
+  def set_salt
+    self.password_salt = BCrypt::Engine.generate_salt if password_salt.blank?
   end
 end
 
@@ -26,10 +26,9 @@ end
 class ChangeLoginSystem < ActiveRecord::Migration
   def up
     execute <<-SQL
-      DELETE FROM users WHERE role='guardian';
+      DELETE FROM users WHERE role='guardian' OR profile_type='AdminProfile';
       UPDATE users SET profile_type='Student' WHERE profile_type='StudentProfile';
       UPDATE users SET profile_type='Teacher' WHERE profile_type='TeacherProfile';
-      UPDATE users SET profile_type='Admin' WHERE profile_type='AdminProfile';
     SQL
 
     change_table :users do |t|
@@ -39,21 +38,21 @@ class ChangeLoginSystem < ActiveRecord::Migration
 
     User.reset_column_information
 
-    n = 1
     User.all.each do |user|
-      user.email = user.profile.email
-      if user.invalid?
-        user.email = "filler_#{n}@email.com"
-        n+= 1
+      if user.profile.email.present?
+        user.email = user.profile.email
+      else
+        user.email = nil
       end
-      user.password = "pass"
       user.save!
     end
 
     Guardian.all.each do |guardian|
-      user = User.new(password: "pass", profile: guardian)
-      user.email = guardian.email
-      user.save! if user.valid?
+      if guardian.email.present?
+        user = User.new(profile: guardian)
+        user.email = guardian.email
+        user.save
+      end
     end
 
     remove_column :students, :email
